feat: Moves to single module in prep for documentation
