--- Sources/AuthController/AuthenticationError.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public enum AuthenticationError: Error {
-  case notAuthenticated
-  case failedToAuthenticate
-}
