+++ Sources/SupabaseClientDependencies/DatabaseRouter/Protocols/RouteCollection.swift
@@ -1,7 +1,8 @@
 import Dependencies
 import PostgREST
 
-/// Represents a router for the database, generally for a single table or collection of routes that are modeled as an enum.
+/// Represents a route collection for the database, generally for a single table or collection of routes that
+/// are modeled as an enum.
 public protocol RouteCollection: Sendable {
 
   /// Return the route container for the given route.
