+++ Tests/SupabaseClientTests/TestModels.swift
@@ -1,5 +1,4 @@
 import CasePaths
-import DatabaseRouter
 import Dependencies
 import Foundation
 import SupabaseClientDependencies
